+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/dao/statistics/ListingStatisticsDAOImpl.java
@@ -10,37 +10,60 @@ import org.springframework.stereotype.Repository;
 import gov.healthit.chpl.dao.impl.BaseDAOImpl;
 import gov.healthit.chpl.domain.DateRange;
 import gov.healthit.chpl.domain.statistics.CertifiedBodyStatistics;
-import gov.healthit.chpl.entity.CertificationStatusType;
 
 @Repository("listingStatisticsDAO")
 public class ListingStatisticsDAOImpl extends BaseDAOImpl implements ListingStatisticsDAO {
     /**
-     * Total # of Certified Unique Products (Regardless of Status or Edition –
-     * Including 2011)
+     * Total # of unique products. Any parameter can be null.
      */
     @Override
-    public Long getTotalCertifiedProducts(DateRange dateRange) {
-        Query query = entityManager.createQuery(
-                "SELECT DISTINCT CONCAT(UPPER(productName), UPPER(developerName)) FROM CertifiedProductDetailsEntity "
-                        + " WHERE (deleted = false AND creationDate BETWEEN :creationStartDate AND :creationEndDate) "
-                        + " OR (deleted = true AND creationDate BETWEEN :creationStartDate AND :creationEndDate AND lastModifiedDate > :creationEndDate) ");
-        query.setParameter("creationStartDate", dateRange.getStartDate());
-        query.setParameter("creationEndDate", dateRange.getEndDate());
-        return (long) query.getResultList().size();
-    }
-
-    /**
-     * Total # of unique Products with 2014 Listings
-     */
-    @Override
-    public Long getTotalCPs2014Listings(DateRange dateRange) {
-        Query query = entityManager.createQuery(
-                "SELECT DISTINCT CONCAT(UPPER(productName), UPPER(developerName)) FROM CertifiedProductDetailsEntity "
-                        + " WHERE year = '2014' "
-                        + " AND (deleted = false AND creationDate BETWEEN :creationStartDate AND :creationEndDate) "
-                        + " OR (deleted = true AND creationDate BETWEEN :creationStartDate AND :creationEndDate AND lastModifiedDate > :creationEndDate)) ");
-        query.setParameter("creationStartDate", dateRange.getStartDate());
-        query.setParameter("creationEndDate", dateRange.getEndDate());
+    public Long getTotalUniqueProductsByEditionAndStatus(DateRange dateRange, String edition, List<String> statuses) {
+        String hql = "SELECT DISTINCT UPPER(productName) || UPPER(developerName) "
+                + "FROM CertifiedProductDetailsEntity ";
+        
+        boolean hasWhere = false;
+        if(edition != null) {
+            hql += " WHERE year = :edition ";
+            hasWhere = true;
+        }
+        if(statuses != null && statuses.size() > 0) {
+            if(!hasWhere) {
+                hql += " WHERE ";
+                hasWhere = true;
+            } else {
+                hql += " AND ";
+            }
+            hql += " UPPER(certificationStatusName) IN (:statuses) ";
+        }
+        if(dateRange == null) {
+            if(!hasWhere) {
+                hql += " WHERE ";
+                hasWhere = true;
+            } else {
+                hql += " AND ";
+            }
+            hql += " deleted = false ";
+        } else {
+            if(!hasWhere) {
+                hql += " WHERE ";
+                hasWhere = true;
+            } else {
+                hql += " AND ";
+            }
+            hql += "((deleted = false AND creationDate <= :endDate) "
+                    + " OR "
+                    + "(deleted = true AND creationDate <= :endDate AND lastModifiedDate > :endDate)) ";
+        }
+        Query query = entityManager.createQuery(hql);
+        if(edition != null) {
+            query.setParameter("edition", edition);
+        }
+        if(statuses != null && statuses.size() > 0) {
+            query.setParameter("statuses", statuses);
+        }
+        if(dateRange != null) {
+            query.setParameter("endDate", dateRange.getEndDate());
+        }
         return (long) query.getResultList().size();
     }
 
@@ -50,16 +73,28 @@ public class ListingStatisticsDAOImpl extends BaseDAOImpl implements ListingStat
      */
     @Override
     public List<CertifiedBodyStatistics> getTotalCPListingsEachYearByCertifiedBody(DateRange dateRange) {
-        String queryStr = "SELECT t.certification_body_name, t.year, count(DISTINCT t.products) "
-                + "FROM(SELECT DISTINCT certification_body_name, year, CONCAT(UPPER(product_Name), UPPER(vendor_Name)) AS products "
-                + "FROM openchpl.certified_product_details "
-                + "WHERE (deleted = false AND creation_Date BETWEEN :creationStartDate AND :creationEndDate) "
-                + "OR (deleted = true AND creation_Date BETWEEN :creationStartDate AND :creationEndDate AND last_Modified_Date > :creationEndDate) "
-                + ") t " + "GROUP BY certification_body_name, year " + "ORDER BY t.certification_body_name ";
+        String sql = "SELECT t.certification_body_name, t.year, count(DISTINCT t.products) "
+                + "FROM( "
+                +   "SELECT DISTINCT certification_body_name, year, CONCAT(UPPER(product_Name), UPPER(vendor_Name)) AS products "
+                +   "FROM openchpl.certified_product_details "
+                +   "WHERE ";
+        if(dateRange == null) {
+            sql += " deleted = false ";
+        } else {
+            sql += " ((deleted = false AND creation_date <= :endDate) "
+                    + " OR "
+                    + "(deleted = true AND creation_date <= :endDate AND last_modified_date > :startDate)) ";
+        }
+        sql += ") t " 
+                + " GROUP BY certification_body_name, year " 
+                + " ORDER BY t.certification_body_name ";
 
-        Query query = entityManager.createNativeQuery(queryStr);
-        query.setParameter("creationStartDate", dateRange.getStartDate());
-        query.setParameter("creationEndDate", dateRange.getEndDate());
+        Query query = entityManager.createNativeQuery(sql);
+        
+        if(dateRange != null) {
+            query.setParameter("startDate", dateRange.getStartDate());
+            query.setParameter("endDate", dateRange.getEndDate());
+        }
 
         List<Object[]> results = query.getResultList();
         List<CertifiedBodyStatistics> cbStats = new ArrayList<CertifiedBodyStatistics>();
@@ -81,17 +116,27 @@ public class ListingStatisticsDAOImpl extends BaseDAOImpl implements ListingStat
     @Override
     public List<CertifiedBodyStatistics> getTotalCPListingsEachYearByCertifiedBodyAndCertificationStatus(
             DateRange dateRange) {
-        String queryStr = "SELECT t.certification_body_name, t.year, count(DISTINCT t.products), t.certification_status_name "
-                + "FROM(SELECT DISTINCT certification_body_name, year, CONCAT(UPPER(product_Name), UPPER(vendor_Name)) AS products, certification_status_name "
-                + "FROM openchpl.certified_product_details "
-                + "WHERE (deleted = false AND creation_Date BETWEEN :creationStartDate AND :creationEndDate) "
-                + "OR (deleted = true AND creation_Date BETWEEN :creationStartDate AND :creationEndDate AND last_Modified_Date > :creationEndDate) "
-                + ") t " + "GROUP BY certification_body_name, year, certification_status_name "
-                + "ORDER BY t.certification_body_name ";
+        String sql = "SELECT t.certification_body_name, t.year, count(DISTINCT t.products), t.certification_status_name "
+                + "FROM( "
+                +   "SELECT DISTINCT certification_body_name, year, CONCAT(UPPER(product_Name), UPPER(vendor_Name)) AS products, certification_status_name "
+                +   "FROM openchpl.certified_product_details "
+                +   "WHERE ";
+        if(dateRange == null) {
+            sql += " deleted = false ";
+        } else {
+            sql += " ((deleted = false AND creation_date <= :endDate) "
+                    + " OR "
+                    + "(deleted = true AND creation_date <= :endDate AND last_modified_date > :startDate)) ";
+        }
+        sql += ") t " 
+                + " GROUP BY certification_body_name, year, certification_status_name "
+                + " ORDER BY t.certification_body_name ";
 
-        Query query = entityManager.createNativeQuery(queryStr);
-        query.setParameter("creationStartDate", dateRange.getStartDate());
-        query.setParameter("creationEndDate", dateRange.getEndDate());
+        Query query = entityManager.createNativeQuery(sql);
+        if(dateRange != null) {
+            query.setParameter("startDate", dateRange.getStartDate());
+            query.setParameter("endDate", dateRange.getEndDate());
+        }
 
         List<Object[]> results = query.getResultList();
         List<CertifiedBodyStatistics> cbStats = new ArrayList<CertifiedBodyStatistics>();
@@ -107,125 +152,57 @@ public class ListingStatisticsDAOImpl extends BaseDAOImpl implements ListingStat
     }
 
     /**
-     * Total # of unique Products with Active 2014 Listings
+     * Total # of Listings. Any parameter can be null.
      */
     @Override
-    public Long getTotalCPsActive2014Listings(DateRange dateRange) {
-        Query query = entityManager.createQuery(
-                "SELECT DISTINCT CONCAT(UPPER(productName), UPPER(developerName)) FROM CertifiedProductDetailsEntity "
-                        + " WHERE year = '2014' AND UPPER(certificationStatusName) = 'ACTIVE' "
-                        + " AND (deleted = false AND creationDate BETWEEN :creationStartDate AND :creationEndDate) "
-                        + " OR (deleted = true AND creationDate BETWEEN :creationStartDate AND :creationEndDate AND lastModifiedDate > :creationEndDate)) ");
-        query.setParameter("creationStartDate", dateRange.getStartDate());
-        query.setParameter("creationEndDate", dateRange.getEndDate());
-        return (long) query.getResultList().size();
-    }
+    public Long getTotalListingsByEditionAndStatus(DateRange dateRange, String edition, List<String> statuses) {
+        String hql = "SELECT COUNT(*) "
+                + "FROM CertifiedProductDetailsEntity ";
+        boolean hasWhere = false;
+        if(edition != null) {
+            hql += " WHERE year = :edition ";
+            hasWhere = true;
+        }
+        if(statuses != null && statuses.size() > 0) {
+            if(!hasWhere) {
+                hql += " WHERE ";
+                hasWhere = true;
+            } else {
+                hql += " AND ";
+            }
+            hql += " UPPER(certificationStatusName) IN (:statuses) ";
+        }
+        if(dateRange == null) {
+            if(!hasWhere) {
+                hql += " WHERE ";
+                hasWhere = true;
+            } else {
+                hql += " AND ";
+            }
+            hql += " deleted = false ";
+        } else {
+            if(!hasWhere) {
+                hql += " WHERE ";
+                hasWhere = true;
+            } else {
+                hql += " AND ";
+            }
+            hql += "((deleted = false AND creationDate <= :endDate) "
+                    + " OR "
+                    + "(deleted = true AND creationDate <= :endDate AND lastModifiedDate > :endDate)) ";
+        }
+        
+        Query query = entityManager.createQuery(hql);
+        if(edition != null) {
+            query.setParameter("edition", edition);
+        }
+        if(statuses != null && statuses.size() > 0) {
+            query.setParameter("statuses", statuses);
+        }
+        if(dateRange != null) {
+            query.setParameter("endDate", dateRange.getEndDate());
+        }
 
-    /**
-     * Total # of unique Products with Suspended (by ONC or ONC-ACB) 2014
-     * Listings
-     */
-    @Override
-    public Long getTotalCPsSuspended2014Listings(DateRange dateRange) {
-        Query query = entityManager.createQuery("SELECT DISTINCT "
-                + " CONCAT(UPPER(productName), UPPER(developerName)) " + " FROM CertifiedProductDetailsEntity "
-                + " WHERE year = '2014' AND UPPER(certificationStatusName) IN ('SUSPENDED BY ONC-ACB', 'SUSPENDED BY ONC') "
-                + " AND (" + "	(deleted = false AND creationDate BETWEEN :creationStartDate AND :creationEndDate) "
-                + " OR " + "	(deleted = true AND creationDate BETWEEN :creationStartDate AND :creationEndDate "
-                + "	AND lastModifiedDate > :creationEndDate)" + ") ");
-        query.setParameter("creationStartDate", dateRange.getStartDate());
-        query.setParameter("creationEndDate", dateRange.getEndDate());
-        return (long) query.getResultList().size();
-    }
-
-    /**
-     * Total # of unique Products with 2015 Listings
-     */
-    @Override
-    public Long getTotalCPs2015Listings(DateRange dateRange) {
-        Query query = entityManager.createQuery("SELECT DISTINCT "
-                + " CONCAT(UPPER(productName), UPPER(developerName)) " + " FROM CertifiedProductDetailsEntity "
-                + " WHERE year = '2015' "
-                + " AND (deleted = false AND creationDate BETWEEN :creationStartDate AND :creationEndDate) "
-                + " OR (deleted = true AND creationDate BETWEEN :creationStartDate AND :creationEndDate AND lastModifiedDate > :creationEndDate)) ");
-        query.setParameter("creationStartDate", dateRange.getStartDate());
-        query.setParameter("creationEndDate", dateRange.getEndDate());
-        return (long) query.getResultList().size();
-    }
-
-    /**
-     * Total # of unique Products with Active 2015 Listings
-     */
-    @Override
-    public Long getTotalCPsActive2015Listings(DateRange dateRange) {
-        Query query = entityManager.createQuery(
-                "SELECT DISTINCT CONCAT(UPPER(productName), UPPER(developerName)) FROM CertifiedProductDetailsEntity "
-                        + " WHERE year = '2015' AND UPPER(certificationStatusName) = 'ACTIVE' "
-                        + " AND (deleted = false AND creationDate BETWEEN :creationStartDate AND :creationEndDate) "
-                        + " OR (deleted = true AND creationDate BETWEEN :creationStartDate AND :creationEndDate AND lastModifiedDate > :creationEndDate)) ");
-        query.setParameter("creationStartDate", dateRange.getStartDate());
-        query.setParameter("creationEndDate", dateRange.getEndDate());
-        return (long) query.getResultList().size();
-    }
-
-    /**
-     * Total # of unique Products with Suspended (by ONC or ONC-ACB) 2015
-     * Listings
-     */
-    @Override
-    public Long getTotalCPsSuspended2015Listings(DateRange dateRange) {
-        Query query = entityManager.createQuery(
-                "SELECT DISTINCT CONCAT(UPPER(productName), UPPER(developerName)) FROM CertifiedProductDetailsEntity "
-                        + " WHERE year = '2015' AND UPPER(certificationStatusName) IN ('SUSPENDED BY ONC-ACB', 'SUSPENDED BY ONC') "
-                        + " AND (deleted = false AND creationDate BETWEEN :creationStartDate AND :creationEndDate) "
-                        + " OR (deleted = true AND creationDate BETWEEN :creationStartDate AND :creationEndDate AND lastModifiedDate > :creationEndDate)) ");
-        query.setParameter("creationStartDate", dateRange.getStartDate());
-        query.setParameter("creationEndDate", dateRange.getEndDate());
-        return (long) query.getResultList().size();
-    }
-
-    /**
-     * Total # of unique Products with Active Listings (Regardless of Edition)
-     */
-    @Override
-    public Long getTotalCPsActiveListings(DateRange dateRange) {
-        Query query = entityManager.createQuery(
-                "SELECT DISTINCT CONCAT(UPPER(productName), UPPER(developerName)) FROM CertifiedProductDetailsEntity "
-                        + " WHERE UPPER(certificationStatusName) = 'ACTIVE' "
-                        + " AND (deleted = false AND creationDate BETWEEN :creationStartDate AND :creationEndDate) "
-                        + " OR (deleted = true AND creationDate BETWEEN :creationStartDate AND :creationEndDate AND lastModifiedDate > :creationEndDate)) ");
-        query.setParameter("creationStartDate", dateRange.getStartDate());
-        query.setParameter("creationEndDate", dateRange.getEndDate());
-        return (long) query.getResultList().size();
-    }
-
-    /**
-     * Total # of Listings (Regardless of Status or Edition)
-     */
-    @Override
-    public Long getTotalListings(DateRange dateRange) {
-        Query query = entityManager.createQuery("SELECT count(*) FROM CertifiedProductDetailsEntity "
-                + " WHERE (deleted = false AND creationDate BETWEEN :creationStartDate AND :creationEndDate) "
-                + " OR (deleted = true AND creationDate BETWEEN :creationStartDate AND :creationEndDate AND lastModifiedDate > :creationEndDate)) ");
-        query.setParameter("creationStartDate", dateRange.getStartDate());
-        query.setParameter("creationEndDate", dateRange.getEndDate());
-        return (Long) query.getSingleResult();
-    }
-
-    /**
-     * Total # of Active (Including Suspended) 2014 Listings
-     */
-    @Override
-    public Long getTotalActive2014Listings(DateRange dateRange) {
-        Query query = entityManager.createQuery("SELECT count(*) FROM CertifiedProductDetailsEntity "
-                + " WHERE year = '2014' AND UPPER(certificationStatusName) IN ('"
-                + CertificationStatusType.Active.getName().toUpperCase() + "', '"
-                + CertificationStatusType.SuspendedByAcb.getName().toUpperCase() + "', '"
-                + CertificationStatusType.SuspendedByOnc.getName().toUpperCase() + "')"
-                + " AND (deleted = false AND creationDate BETWEEN :creationStartDate AND :creationEndDate) "
-                + " OR (deleted = true AND creationDate BETWEEN :creationStartDate AND :creationEndDate AND lastModifiedDate > :creationEndDate)) ");
-        query.setParameter("creationStartDate", dateRange.getStartDate());
-        query.setParameter("creationEndDate", dateRange.getEndDate());
         return (Long) query.getSingleResult();
     }
 
@@ -234,14 +211,26 @@ public class ListingStatisticsDAOImpl extends BaseDAOImpl implements ListingStat
      */
     @Override
     public List<CertifiedBodyStatistics> getTotalActiveListingsByCertifiedBody(DateRange dateRange) {
-        Query query = entityManager
-                .createQuery("SELECT certificationBodyName, year, count(*) FROM CertifiedProductDetailsEntity "
-                        + " WHERE UPPER(certificationStatusName) = 'ACTIVE' "
-                        + " AND (deleted = false AND creationDate BETWEEN :creationStartDate AND :creationEndDate) "
-                        + " OR (deleted = true AND creationDate BETWEEN :creationStartDate AND :creationEndDate AND lastModifiedDate > :creationEndDate) "
-                        + " GROUP BY certificationBodyName, year " + " ORDER BY certificationBodyName ");
-        query.setParameter("creationStartDate", dateRange.getStartDate());
-        query.setParameter("creationEndDate", dateRange.getEndDate());
+        String hql = "SELECT certificationBodyName, year, count(*) "
+                        + "FROM CertifiedProductDetailsEntity "
+                        + "WHERE UPPER(certificationStatusName) = 'ACTIVE' ";
+        if(dateRange == null) {
+            hql += " AND deleted = false ";
+        } else {
+            hql += " AND "
+                    + "((deleted = false AND creationDate <= :endDate) "
+                    + " OR "
+                    + "(deleted = true AND creationDate <= :endDate AND lastModifiedDate > :startDate)) ";
+        }
+        hql += " GROUP BY certificationBodyName, year " 
+                + " ORDER BY certificationBodyName ";
+        
+        Query query = entityManager.createQuery(hql);
+        if(dateRange != null) {
+            query.setParameter("startDate", dateRange.getStartDate());
+            query.setParameter("endDate", dateRange.getEndDate());
+        }
+        
         List<Object[]> results = query.getResultList();
         List<CertifiedBodyStatistics> cbStats = new ArrayList<CertifiedBodyStatistics>();
         for (Object[] obj : results) {
@@ -253,61 +242,5 @@ public class ListingStatisticsDAOImpl extends BaseDAOImpl implements ListingStat
             cbStats.add(stat);
         }
         return cbStats;
-    }
-
-    /**
-     * Total # of Active 2015 Listings
-     */
-    @Override
-    public Long getTotalActive2015Listings(DateRange dateRange) {
-        Query query = entityManager.createQuery("SELECT count(*) FROM CertifiedProductDetailsEntity "
-                + " WHERE year = '2015' AND UPPER(certificationStatusName) IN ('"
-                + CertificationStatusType.Active.getName().toUpperCase() + "', '"
-                + CertificationStatusType.SuspendedByAcb.getName().toUpperCase() + "', '"
-                + CertificationStatusType.SuspendedByOnc.getName().toUpperCase() + "')"
-                + " AND (deleted = false AND creationDate BETWEEN :creationStartDate AND :creationEndDate) "
-                + " OR (deleted = true AND creationDate BETWEEN :creationStartDate AND :creationEndDate AND lastModifiedDate > :creationEndDate)) ");
-        query.setParameter("creationStartDate", dateRange.getStartDate());
-        query.setParameter("creationEndDate", dateRange.getEndDate());
-        return (Long) query.getSingleResult();
-    }
-
-    /**
-     * Total # of 2014 Listings (Regardless of Status)
-     */
-    @Override
-    public Long getTotal2014Listings(DateRange dateRange) {
-        Query query = entityManager.createQuery("SELECT count(*) FROM CertifiedProductDetailsEntity "
-                + " WHERE year = '2014' AND (deleted = false AND creationDate BETWEEN :creationStartDate AND :creationEndDate) "
-                + " OR (deleted = true AND creationDate BETWEEN :creationStartDate AND :creationEndDate AND lastModifiedDate > :creationEndDate)) ");
-        query.setParameter("creationStartDate", dateRange.getStartDate());
-        query.setParameter("creationEndDate", dateRange.getEndDate());
-        return (Long) query.getSingleResult();
-    }
-
-    /**
-     * Total # of 2015 Listings (Regardless of Status)
-     */
-    @Override
-    public Long getTotal2015Listings(DateRange dateRange) {
-        Query query = entityManager.createQuery("SELECT count(*) FROM CertifiedProductDetailsEntity "
-                + " WHERE year = '2015' AND (deleted = false AND creationDate BETWEEN :creationStartDate AND :creationEndDate) "
-                + " OR (deleted = true AND creationDate BETWEEN :creationStartDate AND :creationEndDate AND lastModifiedDate > :creationEndDate)) ");
-        query.setParameter("creationStartDate", dateRange.getStartDate());
-        query.setParameter("creationEndDate", dateRange.getEndDate());
-        return (Long) query.getSingleResult();
-    }
-
-    /**
-     * Total # of 2011 Listings (Will not be active)
-     */
-    @Override
-    public Long getTotal2011Listings(DateRange dateRange) {
-        Query query = entityManager.createQuery("SELECT count(*) FROM CertifiedProductDetailsEntity "
-                + " WHERE year = '2011' AND (deleted = false AND creationDate BETWEEN :creationStartDate AND :creationEndDate) "
-                + " OR (deleted = true AND creationDate BETWEEN :creationStartDate AND :creationEndDate AND lastModifiedDate > :creationEndDate)) ");
-        query.setParameter("creationStartDate", dateRange.getStartDate());
-        query.setParameter("creationEndDate", dateRange.getEndDate());
-        return (Long) query.getSingleResult();
     }
 }