+++ chpl/chpl-service/src/test/java/gov/healthit/chpl/dao/impl/StatisticsDAOTest.java
@@ -1,5 +1,6 @@
 package gov.healthit.chpl.dao.impl;
 
+import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.HashSet;
@@ -43,6 +44,7 @@ import gov.healthit.chpl.domain.SurveillanceRequirementType;
 import gov.healthit.chpl.domain.SurveillanceResultType;
 import gov.healthit.chpl.domain.SurveillanceType;
 import gov.healthit.chpl.domain.statistics.CertifiedBodyStatistics;
+import gov.healthit.chpl.entity.CertificationStatusType;
 import junit.framework.TestCase;
 
 @RunWith(SpringJUnit4ClassRunner.class)
@@ -123,7 +125,7 @@ public class StatisticsDAOTest extends TestCase {
 		DateRange dateRange = new DateRange(new Date(cal.getTimeInMillis()), new Date());
 		Long totalDevelopers = developerStatisticsDao.getTotalDevelopers(dateRange);
 		assertNotNull(totalDevelopers);
-		assertEquals(5L, totalDevelopers.longValue()); 
+		assertEquals(9L, totalDevelopers.longValue()); 
 	}
 	
 	/**
@@ -142,7 +144,7 @@ public class StatisticsDAOTest extends TestCase {
 		DateRange dateRange = new DateRange(new Date(calStart2016.getTimeInMillis()), new Date(calEnd2016.getTimeInMillis()));
 		Long totalDevelopers = developerStatisticsDao.getTotalDevelopers(dateRange);
 		assertNotNull(totalDevelopers);
-		assertEquals(3L, totalDevelopers.longValue()); 
+		assertEquals(7L, totalDevelopers.longValue()); 
 	}
 	
 	/**
@@ -154,7 +156,8 @@ public class StatisticsDAOTest extends TestCase {
 	@Transactional(readOnly = true)
 	public void getTotalDevelopersWith2014Listings_filterDateRange_allDates() {
 		DateRange dateRange = new DateRange(new Date(0), new Date());
-		Long totalDevelopers2014Listings = developerStatisticsDao.getTotalDevelopersWith2014Listings(dateRange);
+		Long totalDevelopers2014Listings = developerStatisticsDao
+		        .getTotalDevelopersWithListingsByEditionAndStatus(dateRange, "2014", null);
 		assertNotNull(totalDevelopers2014Listings);
 		assertEquals(1L, totalDevelopers2014Listings.longValue());
 	}
@@ -168,7 +171,8 @@ public class StatisticsDAOTest extends TestCase {
 	@Transactional(readOnly = true)
 	public void getTotalDevelopersWith2015Listings_filterDateRange_allDates() {
 		DateRange dateRange = new DateRange(new Date(0), new Date());
-		Long totalDevelopers2015Listings = developerStatisticsDao.getTotalDevelopersWith2015Listings(dateRange);
+		Long totalDevelopers2015Listings = developerStatisticsDao
+		        .getTotalDevelopersWithListingsByEditionAndStatus(dateRange, "2015", null);
 		assertNotNull(totalDevelopers2015Listings);
 		assertEquals(3L, totalDevelopers2015Listings.longValue()); 
 	}
@@ -183,7 +187,8 @@ public class StatisticsDAOTest extends TestCase {
 	@Transactional(readOnly = true)
 	public void getTotalCertifiedProducts_filterDateRange_allDates() {
 		DateRange dateRange = new DateRange(new Date(0), new Date());
-		Long totalCertifiedProducts = listingStatisticsDao.getTotalCertifiedProducts(dateRange);
+		Long totalCertifiedProducts = listingStatisticsDao
+		        .getTotalUniqueProductsByEditionAndStatus(dateRange, null, null);
 		assertNotNull(totalCertifiedProducts);
 		assertEquals(8L, totalCertifiedProducts.longValue()); 
 	}
@@ -197,9 +202,12 @@ public class StatisticsDAOTest extends TestCase {
 	@Transactional(readOnly = true)
 	public void getTotalCPsActive2014Listings_filterDateRange_allDates() {
 		DateRange dateRange = new DateRange(new Date(0), new Date());
-		Long totalActive2014Listings = listingStatisticsDao.getTotalCPsActive2014Listings(dateRange);
-		assertNotNull(totalActive2014Listings);
-		assertEquals(1L, totalActive2014Listings.longValue()); 
+		 List<String> activeStatuses = new ArrayList<String>();
+		 activeStatuses.add(CertificationStatusType.Active.getName().toUpperCase());
+		 Long totalActive2014Listings = listingStatisticsDao
+		        .getTotalUniqueProductsByEditionAndStatus(dateRange, "2014", activeStatuses);
+		 assertNotNull(totalActive2014Listings);
+		 assertEquals(1L, totalActive2014Listings.longValue()); 
 	}
 	
 	/**
@@ -211,7 +219,10 @@ public class StatisticsDAOTest extends TestCase {
 	@Transactional(readOnly = true)
 	public void getTotalCPsActive2015Listings_filterDateRange_allDates() {
 		DateRange dateRange = new DateRange(new Date(0), new Date());
-		Long totalActive2015Listings = listingStatisticsDao.getTotalCPsActive2015Listings(dateRange);
+		List<String> activeStatuses = new ArrayList<String>();
+        activeStatuses.add(CertificationStatusType.Active.getName().toUpperCase());
+        Long totalActive2015Listings = listingStatisticsDao
+               .getTotalUniqueProductsByEditionAndStatus(dateRange, "2015", activeStatuses);        
 		assertNotNull(totalActive2015Listings);
 		assertEquals(1L, totalActive2015Listings.longValue()); 
 	}
@@ -226,7 +237,10 @@ public class StatisticsDAOTest extends TestCase {
 	@Transactional(readOnly = true)
 	public void getTotalCPsActiveListings_filterDateRange_allDates() {
 		DateRange dateRange = new DateRange(new Date(0), new Date());
-		Long totalActiveCPs = listingStatisticsDao.getTotalCPsActiveListings(dateRange);
+		List<String> activeStatuses = new ArrayList<String>();
+        activeStatuses.add(CertificationStatusType.Active.getName().toUpperCase());
+        Long totalActiveCPs = listingStatisticsDao
+               .getTotalUniqueProductsByEditionAndStatus(dateRange, null, activeStatuses); 
 		assertNotNull(totalActiveCPs);
 		assertEquals(3L, totalActiveCPs.longValue()); 
 	}
@@ -240,11 +254,44 @@ public class StatisticsDAOTest extends TestCase {
 	@Transactional(readOnly = true)
 	public void getTotalListings_filterDateRange_allDates() {
 		DateRange dateRange = new DateRange(new Date(0), new Date());
-		Long totalListings = listingStatisticsDao.getTotalListings(dateRange);
+        Long totalListings = listingStatisticsDao
+               .getTotalListingsByEditionAndStatus(dateRange, null, null); 
 		assertNotNull(totalListings);
 		assertEquals(16L, totalListings.longValue()); 
 	}
 	
+	@Test
+    @Transactional(readOnly = true)
+    public void getTotalListings_Current() {
+        Long totalListings = listingStatisticsDao
+               .getTotalListingsByEditionAndStatus(null, null, null); 
+        assertNotNull(totalListings);
+        assertEquals(16L, totalListings.longValue()); 
+    }
+	
+	@Test
+    @Transactional(readOnly = true)
+    public void getTotalListings_filterDates_beforeListingDeleted() {
+        Calendar endDate = Calendar.getInstance(TimeZone.getTimeZone("GMT"));
+        endDate.set(2016, 7, 19, 0, 0, 0);
+        DateRange dateRange = new DateRange(new Date(0), endDate.getTime());
+        Long totalListings = listingStatisticsDao
+               .getTotalListingsByEditionAndStatus(dateRange, null, null); 
+        assertNotNull(totalListings);
+        assertEquals(17L, totalListings.longValue()); 
+    }
+	
+	@Test
+    @Transactional(readOnly = true)
+    public void getTotalListings_filterDates_afterListingDeleted() {
+	    Calendar startDate = Calendar.getInstance(TimeZone.getTimeZone("GMT"));
+	    startDate.set(2016, 9, 1, 0, 0, 0);
+        DateRange dateRange = new DateRange(startDate.getTime(), new Date());
+        Long totalListings = listingStatisticsDao
+               .getTotalListingsByEditionAndStatus(dateRange, null, null); 
+        assertNotNull(totalListings);
+        assertEquals(16L, totalListings.longValue()); 
+    }
 	
 	/**
 	 * Given that getTotalActive2014Listings(DateRange) is called 
@@ -255,7 +302,12 @@ public class StatisticsDAOTest extends TestCase {
 	@Transactional(readOnly = true)
 	public void getTotalActive2014Listings_filterDateRange_allDates() {
 		DateRange dateRange = new DateRange(new Date(0), new Date());
-		Long totalActive2014Listings = listingStatisticsDao.getTotalActive2014Listings(dateRange);
+		List<String> activeStatuses = new ArrayList<String>();
+        activeStatuses.add(CertificationStatusType.Active.getName().toUpperCase());
+        activeStatuses.add(CertificationStatusType.SuspendedByAcb.getName().toUpperCase());
+        activeStatuses.add(CertificationStatusType.SuspendedByOnc.getName().toUpperCase());
+		Long totalActive2014Listings = listingStatisticsDao
+	               .getTotalListingsByEditionAndStatus(dateRange, "2014", activeStatuses);		
 		assertNotNull(totalActive2014Listings);
 		assertEquals(2L, totalActive2014Listings.longValue()); 
 	}
@@ -283,7 +335,12 @@ public class StatisticsDAOTest extends TestCase {
 	@Transactional(readOnly = true)
 	public void getTotalActive2015Listings_filterDateRange_allDates() {
 		DateRange dateRange = new DateRange(new Date(0), new Date());
-		Long totalActive2015Listings = listingStatisticsDao.getTotalActive2015Listings(dateRange);
+		List<String> activeStatuses = new ArrayList<String>();
+        activeStatuses.add(CertificationStatusType.Active.getName().toUpperCase());
+        activeStatuses.add(CertificationStatusType.SuspendedByAcb.getName().toUpperCase());
+        activeStatuses.add(CertificationStatusType.SuspendedByOnc.getName().toUpperCase());
+        Long totalActive2015Listings = listingStatisticsDao
+                   .getTotalListingsByEditionAndStatus(dateRange, "2015", activeStatuses);  
 		assertNotNull(totalActive2015Listings);
 		assertEquals(3L, totalActive2015Listings.longValue()); 
 	}
@@ -297,7 +354,8 @@ public class StatisticsDAOTest extends TestCase {
 	@Transactional(readOnly = true)
 	public void getTotal2014Listings_filterDateRange_allDates() {
 		DateRange dateRange = new DateRange(new Date(0), new Date());
-		Long total2014Listings = listingStatisticsDao.getTotal2014Listings(dateRange);
+        Long total2014Listings = listingStatisticsDao
+                   .getTotalListingsByEditionAndStatus(dateRange, "2014", null); 
 		assertNotNull(total2014Listings);
 		assertEquals(3L, total2014Listings.longValue()); 
 	}
@@ -311,7 +369,8 @@ public class StatisticsDAOTest extends TestCase {
 	@Transactional(readOnly = true)
 	public void getTotal2015Listings_filterDateRange_allDates() {
 		DateRange dateRange = new DateRange(new Date(0), new Date());
-		Long total2015Listings = listingStatisticsDao.getTotal2015Listings(dateRange);
+		Long total2015Listings = listingStatisticsDao
+                .getTotalListingsByEditionAndStatus(dateRange, "2015", null); 
 		assertNotNull(total2015Listings);
 		assertEquals(10L, total2015Listings.longValue()); 
 	}
@@ -325,7 +384,8 @@ public class StatisticsDAOTest extends TestCase {
 	@Transactional(readOnly = true)
 	public void getTotal2011Listings_filterDateRange_allDates() {
 		DateRange dateRange = new DateRange(new Date(0), new Date());
-		Long total2011Listings = listingStatisticsDao.getTotal2011Listings(dateRange);
+		Long total2011Listings = listingStatisticsDao
+                .getTotalListingsByEditionAndStatus(dateRange, "2011", null); 
 		assertNotNull(total2011Listings);
 		assertEquals(3L, total2011Listings.longValue()); 
 	}