+++ RELEASE_NOTES.md
@@ -9,6 +9,9 @@ _Date TBD_
 ### Minor features
 * Allow reason for certification status change for any listing. Require reason for certification status change if new status is Withdrawn by ONC-ACB.
 
+### Bugs Fixed
+* Statistics correctly account for deleted listings/unique products.
+
 ---
 
 ## Version 15.3.0